Add DoubleBuffer.h
Former-commit-id: 6ae9d498000cf8488730838790ae39b5912e1a08
