+++ src/ofxNI2.h
@@ -5,15 +5,14 @@
 #include "OpenNI.h"
 #include <assert.h>
 
+#include "DoubleBuffer.h"
+
 namespace ofxNI2
 {
 	void init();
 	
 	class Device;
 	class Stream;
-
-	template <typename PixelType>
-	struct DoubleBuffer;
 	
 	class IrStream;
 	class ColorStream;
@@ -54,42 +53,6 @@ protected:
 	
 	openni::Recorder *recorder;
 };
-
-template <typename PixelType>
-struct ofxNI2::DoubleBuffer
-{
-public:
-	
-	DoubleBuffer() : front_buffer_index(0), back_buffer_index(1), allocated(false) {}
-	
-	void allocate(int w, int h, int channels)
-	{
-		if (allocated) return;
-		allocated = true;
-		
-		pix[0].allocate(w, h, channels);
-		pix[1].allocate(w, h, channels);
-	}
-	
-	PixelType& getFrontBuffer() { return pix[front_buffer_index]; }
-	const PixelType& getFrontBuffer() const { return pix[front_buffer_index]; }
-	
-	PixelType& getBackBuffer() { return pix[back_buffer_index]; }
-	const PixelType& getBackBuffer() const { return pix[back_buffer_index]; }
-	
-	void swap()
-	{
-		std::swap(front_buffer_index, back_buffer_index);
-	}
-	
-private:
-	
-	PixelType pix[2];
-	int front_buffer_index, back_buffer_index;
-	bool allocated;
-
-};
-
 
 class ofxNI2::Stream : public openni::VideoStream::NewFrameListener
 {
